Refactor of enemy statistics and utilties. $14q
EnemyStatistics is it's own MonoBehaviour and contains numerical statistics, including health.
EnemyActions contains utilites like Knockback and Stun, as well as states like isStunned or isAttacking.
Enemybody is reduced to contain centrally available variables for scripts that are not directly on the gameObject like AIManager or gameObjects like the Hitbox or a Rayemitter.