Script Collection, fixes $16c
A static class which can hold references to any main logic script.
The class has a List which hold the objects.
Not intended to hold several instances of an object.

Registering with a Method where you pass in the object.
Retrieving a script with a generic Method where you need to pass in the type.
Removing a script is possible if necessary.