Simple Effects system
Effects can be either triggered on collision or by another script.
Each Effect needs it's own controller script, that must lie on an object with a trigger if onCollision is used.

SoundEffectController for sounds, contains a reference to a StudioEventEmitter from FMOD.
ParticleEffectController for particles, contains a reference to a particle System.

EffectsManager knows all of the effects that are on an object (if the object is more complex, like an enemy for example). Effects can be played through the effects manager.